requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
beartype==0.18.5
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-log==0.4.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==44.0.1
cssselect2==0.7.0
deprecation==2.1.0
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.6
django-cache-memoize==0.2.1
django-cacheops==7.1
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-filter==24.3
django-maintenance-mode==0.21.1
django-redis==5.4.0
django-request-logging==0.7.5
django-rest-framework==0.1.0
django-storages==1.14.5
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docutils==0.21.2
dotty-dict==1.3.1
drf-yasg==1.21.8
ecdsa==0.19.0
et_xmlfile==2.0.0
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==1.3.1
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hiredis==3.1.0
honcho==2.0.0
hpack==4.1.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.1.0
identify==2.6.7
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
iniconfig==2.1.0
invoke==1.7.3
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
keyring==25.6.0
kombu==5.4.2
lxml==5.4.0
more-itertools==10.6.0
multidict==6.1.0
nh3==0.2.20
nodeenv==1.9.1
openpyxl==3.1.5
packaging==24.2
pillow==11.1.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
postgrest==0.10.6
powerline-status==2.7
pre_commit==4.1.0
prompt_toolkit==3.0.50
propcache==0.2.1
psutil==7.0.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.21
pydantic_core==2.27.2
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.10.1
pyphen==0.17.2
pytest==8.3.5
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.1
python-fsutil==0.15.0
python-gitlab==3.15.0
python-jose==3.3.0
python-pptx==1.0.2
python-semantic-release==7.33.2
pytz==2025.1
PyYAML==6.0.2
ragflow-sdk==0.18.0
razorpay==1.4.2
readme_renderer==44.0
realtime==1.0.6
redis==5.2.1
reportlab==4.4.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9
SecretStorage==3.3.3
semver==2.13.0
sentry-sdk==2.22.0
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.5.3
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tomlkit==0.13.2
tqdm==4.67.1
twine==3.8.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
virtualenv==20.29.2
wcwidth==0.2.13
weasyprint==64.0
webencodings==0.5.1
websockets==12.0
wheel==0.45.1
whitenoise==6.9.0
XlsxWriter==3.2.3
yarl==1.18.3
zipp==3.21.0
zopfli==0.2.3.post1
